#137 Invoke Dexcom functions when transitioning from 'settings screen' to 'insights screen'.
Invoke Dexcom functions when transitioning from 'settings screen' to 'insights screen'.
